+++ src/components/ui/Chat/ChannelList/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { HTMLAttributes } from 'react';
+import React, { FC, HTMLAttributes } from 'react';
 
 import { ChannelListContext } from './ChannelListContext';
 import { StyledChannelList } from './Styled';
@@ -10,14 +10,14 @@ import { BaseProps } from '../../Base';
 export interface ChannelListProps
   extends Omit<HTMLAttributes<HTMLUListElement & HTMLLIElement>, 'css'>,
     BaseProps {
-      activeChannelArn: string;
-      setActiveChannelArn: (arn: string) => void;
+      activeChannelId: string;
+      setActiveChannelId: (id: string) => void;
 }
 
-export const ChannelList = (props: ChannelListProps) => {
-  const { activeChannelArn, setActiveChannelArn } = props;
+export const ChannelList: FC<ChannelListProps> = (props) => {
+  const { activeChannelId, setActiveChannelId } = props;
 
-  const channelListContext = { setActiveChannelArn, activeChannelArn };
+  const channelListContext = { setActiveChannelId, activeChannelId };
 
   return (
     <ChannelListContext.Provider value={channelListContext}>