+++ src/components/ui/Chat/ChatBubble/Styled.tsx
@@ -3,12 +3,12 @@
 
 import styled from 'styled-components';
 
-import { baseStyles, baseSpacing, BaseProps } from '../../Base';
-import { MessageProps } from '../Message';
+import { baseStyles, baseSpacing } from '../../Base';
+import { ChatBubbleProps } from '.';
 
-interface StyledMessageProps extends MessageProps {}
+interface StyledChatBubbleProps extends ChatBubbleProps {}
 
-export const StyledMessage = styled.div<StyledMessageProps>`
+export const StyledChatBubble = styled.div<StyledChatBubbleProps>`
   display: flex;
   flex-direction: row;
   margin: 0.5rem 0;
@@ -20,17 +20,18 @@ export const StyledMessage = styled.div<StyledMessageProps>`
   ${baseStyles}
 `;
 
-export const StyledMessageContent = styled.div<StyledMessageProps>`
-  background-color: ${props => props.theme.chatMessage[props.variant].bgd};
+export const StyledChatBubbleContent = styled.div<StyledChatBubbleProps>`
+  background-color: ${props => props.theme.chatBubble[props.variant].bgd};
   padding: 0.625rem 1rem;
   border-radius: 4px;
   width: fit-content;
-  color: ${props => props.theme.chatMessage[props.variant].fontColor};
+  color: ${props => props.theme.chatBubble[props.variant].fontColor};
   box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.1);
   line-height: 20px;
   width: fit-content;
   max-width: 70.6%;
   font-size: 0.875rem;
+  position: relative;
 
   & .ch-sender-name {
     font-weight: bold;
@@ -38,30 +39,30 @@ export const StyledMessageContent = styled.div<StyledMessageProps>`
   }
 
   & svg {
+    position: absolute;
     height: 9px;
     width: 4px;
-    position: absolute;
     margin-top: -3px;
     margin-left: -20px;
 
     & .ch-chat-caret {
-      fill: ${props => props.theme.chatMessage[props.variant].bgd}
+      fill: ${props => props.theme.chatBubble[props.variant].bgd}
     }
   }
 `;
 
-export const StyledMessageInfo = styled.div`
+export const StyledChatBubbleInfo = styled.div`
   display: flex;
   margin-right: 0.5rem;
   margin-left: auto;
-  color: ${props => props.theme.chatMessage.container.fontColor};
+  color: ${props => props.theme.chatBubble.container.fontColor};
   align-items: center;
 
-  & .ch-more-options {
+  & .ch-message-actions {
     border-radius: 50%;
   }
 
-  & button:hover .ch-more-options {
+  & button:hover .ch-message-actions {
     background-color: ${props => props.theme.buttons.icon.hover.bgd};
 
     & g {
@@ -69,16 +70,16 @@ export const StyledMessageInfo = styled.div`
     }
   }
 
-  & button:focus .ch-more-options {
+  & button:focus .ch-message-actions {
     border: 1px solid ${props => props.theme.colors.primary.dark};
     border-radius: 50%;
   }
 
-  & .ch-more-options g {
-    fill: ${props => props.theme.chatMessage.container.fontColor};
+  & .ch-message-actions g {
+    fill: ${props => props.theme.chatBubble.container.fontColor};
   }
 
-  & .ch-more-options.isOpen {
+  & .ch-message-actions.isOpen {
     background-color: ${props => props.theme.buttons.icon.active.bgd};
     border-radius: 50%;