+++ src/components/ui/Chat/ChatBubble/index.tsx
@@ -1,0 +1,67 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import React, { HTMLAttributes, ReactNode } from 'react';
+
+import { BaseProps } from '../../Base';
+import { StyledChatBubble, StyledChatBubbleInfo, StyledChatBubbleContent } from './Styled';
+import PopOver from '../../PopOver';
+import { Dots } from '../../icons';
+
+export type MessageVariant = 'outgoing' | 'incoming';
+
+export interface ChatBubbleProps extends Omit<HTMLAttributes<HTMLDivElement>, 'css'>, BaseProps {
+    /* Determines styling for outgoing and incoming messages. */
+    variant: MessageVariant;
+    /* The name of the user that sent the message. */
+    senderName: string;
+    /* The time the message was originally sent. */
+    createdTimestamp: string;
+    /* The time the message was last edited */
+    lastUpdatedTimestamp?: string;
+    /* The text content of the message. */
+    content: string;
+    /* This allows for custom elements to be rendered into the optional popover menu for each message, such as an edit button. */
+    actions?: ReactNode | ReactNode[];
+    /* Determines if the name should be shown or not. */
+    showName?: boolean;
+    /* Adds the bubble tail style to a message. */
+    showTail?: boolean;
+    /* Determines if the message has been removed after being sent. */
+    redacted?: boolean;
+    /* Determines if the message has been edited after being sent. */
+    edited?: boolean;
+}
+
+const ChatBubble = ({ showName = true, ...props }: ChatBubbleProps) => {
+    return (
+        <StyledChatBubble {...props} data-testid='chat-bubble'>
+            <StyledChatBubbleContent {...props}>
+                {showName && <div className='ch-sender-name'>{props.senderName}</div>}
+                <div>{props.content}</div>
+                {props.showTail &&   
+                <svg viewBox="0 0 4 9">
+                    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+                      <g transform="translate(0.000000, -53.000000)" className='ch-chat-caret'>
+                        <path d="M4,62 L3.92789928,61.999999 C2.89671177,62.0004988 1.33197354,61.8123902 0.200755581,60.8250184 C-0.0781312712,60.5814641 -0.0639788041,60.0290387 0.229060515,59.8181428 C1.47198013,58.9247413 3.99237825,57.6821586 4,52.9112516 L4,52.9112516 L4,62 Z" />
+                      </g>
+                    </g>
+                </svg>}
+            </StyledChatBubbleContent>
+            <StyledChatBubbleInfo>
+                <span>{props.createdTimestamp}</span>
+                {props.actions &&
+                <PopOver
+                    a11yLabel="Open channel options"
+                    placement="bottom-end"
+                    renderButton={isOpen => (
+                        <Dots width='2rem' height='2rem' className={`${isOpen ? 'isOpen' : ''} ch-message-actions`} />
+                    )}
+                    children={props.actions}
+                />}
+            </StyledChatBubbleInfo>
+        </StyledChatBubble>
+    )
+}
+
+export default ChatBubble;