+++ src/components/ui/Chat/DateHeader/Styled.tsx
@@ -1,0 +1,23 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import styled from 'styled-components';
+
+import { baseStyles, baseSpacing } from '../../Base';
+import { DateHeaderProps } from '.';
+
+interface StyledDateHeaderProps extends DateHeaderProps {}
+
+export const StyledDateHeader = styled.div<StyledDateHeaderProps>`
+  border-radius: 13px;
+  background-color: ${props => props.theme.chatDateHeader.bgd};
+  color: ${props => props.theme.chatDateHeader.fontColor};
+  font-size: ${props => props.theme.fontSizes.footnote.fontSize};
+  width: fit-content;
+  padding: 0.25rem;
+  margin: 0 auto;
+
+  ${baseStyles};
+  ${baseSpacing};
+`;
+