+++ src/components/ui/Chat/InfiniteList/index.tsx
@@ -1,39 +1,35 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { useRef, useEffect, useState, HTMLAttributes } from 'react';
+import React, { useRef, useEffect, HTMLAttributes, ComponentType, ReactNode } from 'react';
 
 import { StyledInfiniteList } from './Styled';
-import { BaseProps } from '../Base';
-import { Spinner } from '../icons'; // TODO: update icon
+import { BaseProps } from '../../Base';
+import { Spinner } from '../../icons';
 
 export interface InfiniteListProps extends Omit<HTMLAttributes<HTMLUListElement>, 'css'>, BaseProps {
-  /* The elements that will populat the list. These elements will be wrapped in <li> tags, so that can be any element. */
-  items: JSX.Element[];
   /* A callback function that will make an api call to load the next batch of items. */
   onLoad: () => void;
-  /* Accessibility label of the list */
-  label: string;
-  /* Manages the visibility of the spinner when the API call is resolving */
+  /* Manages the visibility of the spinner when the API call is resolving. */
   isLoading: boolean;
+  items: ReactNode[];
 }
-
 const InfiniteList = (props: InfiniteListProps) => {
-  const { items, onLoad, label, isLoading } = props;
+  const { isLoading, onLoad, items } = props;
 
-  const feedRef = useRef<HTMLUListElement>(null);
+  const containerRef = useRef<HTMLUListElement>(null);
   const listEnd = useRef<HTMLDivElement>(null);
-  const currentTopItem = useRef<HTMLLIElement>(null);
+  const currentTopItemRef = useRef<HTMLLIElement>(null);
   const firstNew = useRef<HTMLLIElement>(null);
   const prevLength = useRef(items.length)
   const newLength = useRef(0);
-
   const onLoadRef = useRef(onLoad);
+
   onLoadRef.current = onLoad;
 
   useEffect(() => {
     firstNew.current?.scrollIntoView();
-  }, [items.length])
+  }, [items.length]);
 
   useEffect(() => {
     listEnd.current?.scrollIntoView();
@@ -44,26 +40,25 @@ const InfiniteList = (props: InfiniteListProps) => {
       }
     },
       { 
-        root: feedRef.current,
+        root: containerRef.current,
         threshold: 1
       }
     );
     
-    if (currentTopItem.current) {
-      observer.observe(currentTopItem.current);
+    if (currentTopItemRef.current) {
+      observer.observe(currentTopItemRef.current);
     }
     return () => {
-      if (currentTopItem.current) {
-        observer.unobserve(currentTopItem.current)
+      if (currentTopItemRef.current) {
+        observer.unobserve(currentTopItemRef.current)
       }
     }
-  }, [])
+  }, []);
 
   if (items.length !== prevLength.current) {
     prevLength.current = newLength.current;
   }
   newLength.current = items.length;
-
   const getRef = (index: number) => {
     if (
       index === (items.length - prevLength.current) - 1 &&
@@ -76,23 +71,19 @@ const InfiniteList = (props: InfiniteListProps) => {
     }
   };
 
+  const messageList = items.map((item: JSX.Element, i: number) => <li id={i.toString()} key={i} ref={i === 0 ? currentTopItemRef : getRef(i)} role='article'>{item}</li>);
+
   return (
     <StyledInfiniteList
-      ref={feedRef}
-      className={isLoading ? 'ch-not-scrollable' : ''}
       {...props}
+      ref={containerRef}
+      className={`${isLoading ? 'ch-not-scrollable' : ''} infinite`}
       data-testid='infinite-list'
       aria-busy={isLoading ? true : false}
       role='feed'
-      aria-label={label}
     >
       {isLoading && <li className="ch-spinner"><Spinner /></li>}
-      {items.map((item, index) => {
-        if (index === 0) {
-          return <li id={index.toString()} key={index} ref={currentTopItem} role='article'>{item}</li>
-        }
-        return <li id={index.toString()} key={index} ref={getRef(index)} role='article'>{item}</li>
-      })}
+      {messageList}
       <div ref={listEnd} />
     </StyledInfiniteList>
   )