+++ src/components/ui/ControlBar/ControlBarItem.tsx
@@ -1,17 +1,17 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { FC } from 'react';
+import React, { FC, ReactNode } from 'react';
 
 import Caret from '../icons/Caret';
 import { StyledControlBarItem, isVertical } from './Styled';
-import PopOver from '../PopOver';
+import PopOver, { Placement } from '../PopOver';
 import PopOverItem, { PopOverItemProps } from '../PopOver/PopOverItem';
 import { useControlBarContext } from './ControlBarContext';
 import IconButton from '../Button/IconButton';
 import { BaseProps } from '../Base';
 
-export interface ControlBarItemProps
+export interface ControlBarButtonProps
   extends Omit<React.HTMLAttributes<HTMLDivElement>, 'css'>,
     BaseProps {
   /** The icon of the control bar item. */
@@ -22,16 +22,22 @@ export interface ControlBarItemProps
   label: string;
   /** The items to render in a popover menu. When passed, the button will render an arrow to open or close a popover menu. Refer to [PopOverItem](/?path=/docs/ui-components-popover--basic-pop-over-menu) */
   popOver?: PopOverItemProps[] | null;
+  /** Defines the placement of PopOver menu. */
+  popOverPlacement?: Placement;
   /**  Apply this prop to receive visual feedback that the button is 'active' */
   isSelected?: boolean;
+  /** Use children to define an alternative to popOver prop with a custom set of elements to be rendered into the popover */
+  children?: ReactNode | ReactNode[];
 }
 
-export const ControlBarItem: FC<ControlBarItemProps> = ({
+export const ControlBarButton: FC<ControlBarButtonProps> = ({
   icon,
   onClick,
   label,
   isSelected = false,
   popOver = null,
+  popOverPlacement,
+  children,
   ...rest
 }) => {
   const context = useControlBarContext();
@@ -40,11 +46,14 @@ export const ControlBarItem: FC<ControlBarItemProps> = ({
     <PopOver
       renderButton={isOpen => getButtonContents(isOpen)}
       a11yLabel={label}
-      children={popOver?.map((option: PopOverItemProps, index: number) => (
+      className="ch-control-bar-popover"
+      placement={popOverPlacement}
+    >
+      {popOver?.map((option: PopOverItemProps, index: number) => (
         <PopOverItem {...option} key={index} />
       ))}
-      className="ch-control-bar-popover"
-    />
+      {children}
+    </PopOver>
   );
 
   const getButtonContents = (isOpen: boolean) => (
@@ -64,7 +73,7 @@ export const ControlBarItem: FC<ControlBarItemProps> = ({
       popOver={popOver}
     >
       <IconButton onClick={onClick} label={label} icon={icon} className="ch-control-bar-item-iconButton"/>
-      {!!popOver && renderPopOver()}
+      {(popOver || children) && renderPopOver()}
       {context.showLabels && (
         <div className="ch-control-bar-item-label">{label}</div>
       )}
@@ -72,4 +81,4 @@ export const ControlBarItem: FC<ControlBarItemProps> = ({
   );
 };
 
-export default ControlBarItem;
+export default ControlBarButton;