+++ src/components/ui/Utilities/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { formatTime } from './formatTime';
+export { formatDate,  DateOptions } from './formatDate';