+++ src/index.ts
@@ -9,7 +9,7 @@ export { SecondaryButton } from './components/ui/Button/SecondaryButton';
 export { IconButton } from './components/ui/Button/IconButton';
 export { Checkbox } from './components/ui/Checkbox';
 export { ControlBar } from './components/ui/ControlBar';
-export { ControlBarItem } from './components/ui/ControlBar/ControlBarItem';
+export { ControlBarButton } from './components/ui/ControlBar/ControlBarItem';
 export { Flex } from './components/ui/Flex';
 export { FormField } from './components/ui/FormField';
 export { Heading } from './components/ui/Heading';
@@ -47,6 +47,8 @@ export { RosterHeader } from './components/ui/Roster/RosterHeader';
 export { RosterGroup } from './components/ui/Roster/RosterGroup';
 export { RosterCell } from './components/ui/Roster/RosterCell';
 export { UserActivityManager } from './components/ui/UserActivityManager';
+export { ChannelList } from './components/ui/Chat/ChannelList';
+export { ChannelItem } from './components/ui/Chat/ChannelList/ChannelItem'
 
 // SDK components
 export {