+++ tst/snapshots/Chat/ChannelList/index.test.tsx
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 describe('ChannelList', () => {
-    it('basic view', async () => {
-      await page.goto('http://host.docker.internal:9009/iframe.html?id=ui-components-channellist--channel-list&viewMode=story');
-      const image = await page.screenshot();
-      expect(image).toMatchImageSnapshot();
-    });
-});  
+  it('ChannelList', async () => {
+    await page.goto('http://host.docker.internal:9009/iframe.html?id=ui-components-chat--channel-list&viewMode=story');
+    const image = await page.screenshot();
+    expect(image).toMatchImageSnapshot();
+  });
+});